+++ extensions/elgato-key-light/src/increaseBrightness.ts
@@ -1,12 +1,13 @@
 import { showToast, Toast } from "@raycast/api";
 import { showFailureToast } from "@raycast/utils";
-import { discoverKeyLights } from "./utils";
+import { discoverKeyLights, getTargetLightNames } from "./utils";
 
 const command = async () => {
   try {
     const keyLight = await discoverKeyLights();
+    const targets = await getTargetLightNames();
     try {
-      const brightness = await keyLight.increaseBrightness();
+      const brightness = await keyLight.increaseBrightness(targets);
 
       await showToast({
         style: Toast.Style.Success,